requirements.txt
# Runtime
pandas>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.18.0
webdriver-manager>=4.0.1
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
schedule>=1.2.1
lxml>=5.1.0
fake-useragent>=1.4.0

# Dev / Quality
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.2.0
ruff>=0.3.0
